feat: migrate to Spring AI 2.0

+++ src/main/java/org/springframework/ai/zhipuai/util/ApiUtils.java
@@ -4,9 +4,9 @@ public class ApiUtils {
 
     public static final Integer DEFAULT_MAX_TOKENS = 1024;
 
-    public static final Float DEFAULT_TEMPERATURE = 0.95f;
+    public static final Double DEFAULT_TEMPERATURE = 0.95D;
 
-    public static final Float DEFAULT_TOP_P = 0.7f;
+    public static final Double DEFAULT_TOP_P = 0.7D;
 
 
 }
